--- src/main/java/chess/application/ChessApplication.java
@@ -1,9 +1,0 @@
-package chess.application;
-
-import chess.controller.ChessController;
-
-public class ChessApplication {
-    public static void main(String[] args) {
-        new ChessController().run();
-    }
-}
