--- src/main/java/chess/domain/piece/Bishop.java
@@ -1,8 +1,0 @@
-package chess.domain.piece;
-
-public class Bishop extends SliderPiece {
-
-    public Bishop(final Color color) {
-        super(color);
-    }
-}
