--- src/main/java/chess/domain/piece/Queen.java
@@ -1,8 +1,0 @@
-package chess.domain.piece;
-
-public class Queen extends SliderPiece {
-
-    public Queen(final Color color) {
-        super(color);
-    }
-}
