--- src/main/java/chess/domain/piece/Rook.java
@@ -1,8 +1,0 @@
-package chess.domain.piece;
-
-public class Rook extends SliderPiece {
-
-    public Rook(final Color color) {
-        super(color);
-    }
-}
