+++ src/main/java/chess/domain/piece/slider/Bishop.java
@@ -1,0 +1,10 @@
+package chess.domain.piece.slider;
+
+import chess.domain.piece.Color;
+
+public class Bishop extends SliderPiece {
+
+    public Bishop(final Color color) {
+        super(color);
+    }
+}
