+++ src/main/java/chess/domain/piece/slider/Queen.java
@@ -1,0 +1,10 @@
+package chess.domain.piece.slider;
+
+import chess.domain.piece.Color;
+
+public class Queen extends SliderPiece {
+
+    public Queen(final Color color) {
+        super(color);
+    }
+}
