+++ src/main/java/chess/domain/piece/slider/Rook.java
@@ -1,0 +1,10 @@
+package chess.domain.piece.slider;
+
+import chess.domain.piece.Color;
+
+public class Rook extends SliderPiece {
+
+    public Rook(final Color color) {
+        super(color);
+    }
+}
