+++ src/test/java/chess/domain/board/BoardTest.java
@@ -1,5 +1,6 @@
 package chess.domain.board;
 
+import chess.domain.board.factory.BoardFactory;
 import chess.domain.board.position.Position;
 import chess.domain.piece.Color;
 import org.junit.jupiter.api.BeforeEach;
