+++ src/test/java/chess/helper/BoardFixture.java
@@ -1,14 +1,14 @@
 package chess.helper;
 
 import chess.domain.board.position.Position;
-import chess.domain.piece.Bishop;
+import chess.domain.piece.slider.Bishop;
 import chess.domain.piece.Color;
-import chess.domain.piece.King;
-import chess.domain.piece.Knight;
-import chess.domain.piece.Pawn;
+import chess.domain.piece.jumper.King;
+import chess.domain.piece.jumper.Knight;
+import chess.domain.piece.pawn.Pawn;
 import chess.domain.piece.Piece;
-import chess.domain.piece.Queen;
-import chess.domain.piece.Rook;
+import chess.domain.piece.slider.Queen;
+import chess.domain.piece.slider.Rook;
 
 import java.util.HashMap;
 import java.util.Map;
